requirements.txt
pyyaml
simplejson
terminaltables
openstacksdk
